pyproject.toml
[tool.black]
line-length = 120

[tool.poetry]
name = "unmess-service"
version = "0.1.0"
description = "Unmess Service"
authors = ["Ashish Kapoor <ashish@sannty.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0.7"
djangorestframework = "3.15.2"
environs = "11.0.0"
django-cors-headers = "4.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"